+++ src/lib/IFrameCommunicatorTypes.ts
@@ -1,4 +1,5 @@
 interface ISDKSupportedFeatures {
+  readonly previewIFrameCurrentUrlHandler: boolean;
   readonly refreshHandler: boolean;
 }
 
@@ -43,6 +44,10 @@ export interface IRefreshMessageData {
 
 export interface IRefreshMessageMetadata {
   readonly manualRefresh: boolean;
+}
+
+export interface IPreviewIFrameCurrentUrlMessageData {
+  readonly previewUrl: string;
 }
 
 export enum InsertPositionPlacement {
@@ -117,6 +122,8 @@ export enum IFrameMessageType {
   AddInitial = 'kontent-smart-link:add:initial',
   AddAction = 'kontent-smart-link:add:action',
   RefreshPreview = 'kontent-smart-link:preview:refresh',
+  PreviewIFrameCurrentUrl = 'kontent-smart-link:preview:current-url',
+  PreviewIFrameCurrentUrlResponse = 'kontent-smart-link:preview:current-url:response',
 }
 
 export enum IFrameResponseType {