+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ability to get currently displayed preview URL from the iframe.
+- New iframe messages added to iframe communication.
+- New message handler.
+
 ## [3.0.0] - 2022-09-23
 
 - Package renamed to `@kontent-ai/smart-link`.